Complete overhaul of MIME API and testing tool to deal with several parsing and decoding bugs.
The issues addressed by this major revision relate to a change
in design assumptions made in previous builds by the decoupling
of source and decode buffers. As a result several key points of
RFC 2046 were no longer handled correctly.

In particular RFC 2046 section 5.1.1

	The boundary delimiter line is then defined as a line
	consisting entirely of two hyphen characters ("-",
	decimal value 45) followed by the boundary parameter
	value from the Content-Type header field, optional
	linear whitespace, and a terminating CRLF.

	...

	The boundary delimiter MUST occur at the beginning of a
	line, i.e., following a CRLF, and the initial CRLF is
	considered to be attached to the boundary delimiter line
	rather than part of the preceding part.  The boundary
	may be followed by zero or more characters of linear
	whitespace. It is then terminated by either another CRLF
	and the header fields for the next part, or by two
	CRLFs, in which case there are no header fields for the
	next part.  If no Content-Type field is present it is
	assumed to be "message/rfc822" in a "multipart/digest"
	and "text/plain" otherwise.

	...

	boundary := 0*69<bchars> bcharsnospace

	bchars := bcharsnospace / " "

	bcharsnospace := DIGIT / ALPHA / "'" / "(" / ")" /
		      "+" / "_" / "," / "-" / "." /
		      "/" / ":" / "=" / "?"

The lazy MIME boundary parsing excludes a boundary marker that
is of zero length or starts with whitespace, eg.

	Content-Type: multipart/mixed; boundary=""

	Content-Type: multipart/mixed; boundary=" foo bar"

Ignoring zero length boundary markers ensures that email
signatures, often delimited by a simple "--CRLF" (see
Thunderbird MUA) or "-- CRLF" remain part of the content.

Thus the following special cases had to be addressed carefully:

a) Care with quoted-printable split line of hyphens, with a
soft-line break, eg.

	\r\n
	-----------------------------------=\r\n
	-------\r\n

These should not be treated as two boundaries.

b) Care for signature line separator that should not be treated
as a boundary even though it is valid, even when using quoted-
printable.

	-- \r\n
or
	--=20\r\n

c) Care with a valid boundary line that looks like a quoted
printable soft line break and with MIME parts that are
quoted-printable encoded.

	\r\n
	--=__PartF0D8505D.0__=\r\n

The mime-eicar-test.eml is crafted to test all of these special
cases using the mime CLI tool.