Linux: Work-around for epoll_wait's lack of pthread cancellation point.
In smtpf at time of a clean shutdown (serverStop), worker
threads may be blocked in the epoll implementation used in
socketTimeouts(). pthread_cancel() does not appear to affect the
behaviour of epoll_wait() and online searches show some stupid
reluctance on the part of the glibc maintainer to fix it.

The solution in serverWorkerCancel() is to call pthread_kill()
immediately after pthread_cancel() to send SIGUSR1 to a thread,
which invokes a no-op signal handler. epoll_wait() and any other
IO functions, that have no cancellation points, to unblock and
set errno to EINTR. The cancellation will then be caught at the
next cancellation point or explicit pthread_testcancel() check.