Refactored the MIME API to support a linked-list of MimeHooks. This is to allow one Mime object to be used by many different APIs at once and so avoid duplicate parsing and decoding of a message. The Mime structure and setup have changed so will require products using the MIME API to make some minor changes.
Add -m option to the mime CLI to generate MD5 for the encoded and
decoded MIME part as proof of concept.

The above MIME API change found three bug: the body_finish
handler was not called on EOF; when a MIME part does not end in
a CRLF, which is part of the boundary, then the decoded portion
should also not end in a CRLF; the mime_part_length should be
reset between parts.