smtpe has replaced the Lua scope model with a master state and a Lua thread pair per SMTP connection. This means the Lua script is reloaded for each new SMTP connection. In this way each connection has its own "global" scope.
The previous smtpe Lua scoping model, a single master state from
which a Lua thread per SMTP connection would be created. "globals"
create by a thread were to be confined to the thread's scope they
were created in, while globals create when the master state was
loaded would be accessible by all threads. This failed to work
and/or required to much hoop jumping, since Lua functions receive
an environment when they are defined (because of closures), not
from the runtime ie. functions do not inherit the environment of
the caller. Essentially confusing to make work correctly.

Also fixed use of old incorrect constants used for eventsWait();
added events-wait option to allow runtime testing of kqueue, epoll,
and poll variants.